+++ NEWS.md
@@ -3,6 +3,7 @@
 - updating dependencies and required versions
 - fix vignette due to updates in `ptable`-pkg
 - simplify examples by using examplary ptable from `ptable`-pkg using `ptable::pt_ex_cnts()` and `ptable::pt_ex_nums()`
+- feature: allow tabulation of non-perturbed variables in `freqtab()`
 
 # cellKey 0.18.0
 - allow to save perturbation-schemes for different variables in `params_cnts_set()` and `params_nums_set()`